requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
biopython==1.84 